+++ internal/ast/ast.go
@@ -232,7 +232,7 @@ var OpTable = []opTableEntry{
 }
 
 type Loc struct {
-	// This is the 0-based index of this location from the start of the file
+	// This is the 0-based index of this location from the start of the file, in bytes
 	Start int32
 }